+++ test/resources/ns2_cleaned.clj
@@ -1,8 +1,8 @@
 (ns resources.ns2-cleaned
   (:require
    [clojure
-    [edn :refer :all :rename {read-string rs read rd}]
+    [edn :refer :all :rename {read-string rs}]
     [instant :refer :all]
     [pprint :refer [cl-format fresh-line get-pretty-writer]]
-    [string :refer :all :rename {replace foo reverse bar}]
+    [string :refer :all]
     [test :refer :all]]))